--- frontend/src/components/BurgerMenu/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './BurgerMenu';
