+++ frontend/src/components/Header/BurgerMenu/__tests__/BurgerMenu.test.tsx
@@ -1,6 +1,6 @@
 import { render } from 'services/testing/reactTestingLibraryWrapper';
 
-import BurgerMenu from '../';
+import { BurgerMenu } from '../';
 
 test('AAU, I can see a BurgerMenu', () => {
   const component = render(<BurgerMenu title="Test" sections={['Section 1', 'Section 2']} />);
