+++ frontend/src/components/Header/BurgerMenu/index.tsx
@@ -1,0 +1,1 @@
+export { BurgerMenu } from './BurgerMenu';
